Moved interface ButtonState to test code
